Move domain object files to respective component folders
