+++ examples/bpi-monolith-test-suite/bpi/components/identity/bpiSubject.ts
