+++ examples/bpi-monolith-test-suite/bpi/components/identity/identity.ts
@@ -1,4 +1,4 @@
-import { BpiSubject } from "../../bpiSubject";
+import { BpiSubject } from "./bpiSubject";
 import { IIdentityComponent } from "./identity.interface";
 
 export class IdentityComponent implements IIdentityComponent {
