+++ examples/bpi-monolith-test-suite/bpi/components/messaging/bpiMessage.ts
@@ -1,4 +1,4 @@
-import { BpiSubject } from "./bpiSubject";
+import { BpiSubject } from "../identity/bpiSubject";
 
 export class BpiMessage {
     id: string;
