+++ examples/bpi-monolith-test-suite/bpi/components/storage/storage.interface.ts
@@ -1,4 +1,4 @@
-import { Agreement } from "../../agreement";
+import { Agreement } from "./agreement";
 
 export interface IStorageComponent {
     setInitialAgreementState(agreement: Agreement): void;
