+++ examples/bpi-monolith-test-suite/bpi/components/storage/storage.ts
@@ -1,4 +1,4 @@
-import { Agreement } from "../../agreement";
+import { Agreement } from "./agreement";
 import { IStorageComponent } from "./storage.interface";
 
 export class StorageComponent implements IStorageComponent {
