+++ examples/bpi-monolith-test-suite/bpi/components/workgroup/workgroup.interface.ts
@@ -1,8 +1,8 @@
 import { Workstep } from '../../workstep';
-import { Workgroup } from '../../workgroup';
-import { BpiSubject } from '../../bpiSubject';
-import { Agreement } from '../../agreement';
-import { Invitation } from '../../invitation';
+import { Workgroup } from './workgroup';
+import { BpiSubject } from '../identity/bpiSubject';
+import { Agreement } from '../storage/agreement';
+import { Invitation } from './invitation';
 
 
 export interface IWorkGroupComponent {
